backend/requirements_pymysql.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.0
